Submission/frontend_service/requirements.txt
fastapi
psycopg2-binary
pytest
python-dotenv
SQLAlchemy
uvicorn

